backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pinecone-client==3.0.0
langchain==0.1.0
langchain-openai==0.0.11
langchain-anthropic==0.1.0
langgraph==0.0.32
langfuse==2.27.0
openai==1.3.0
requests==2.31.0
pdf2image==1.16.3
python-pptx==0.6.21
python-docx==0.8.11
PyPDF2==3.0.1
pillow==10.1.0
torch==2.1.1
transformers==4.35.2
sentence-transformers==2.2.2
numpy==1.26.2
pandas==2.1.3
